explorer/requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
celery-with-redis==3.0
Django==1.6.5
django-celery==3.1.16
django-celery-email-with-retry==1.0.3
django-haystack==2.1.0
django-redis==3.8.3
django-uuidfield==0.5.0
django-email-log==0.2.0
enum34==0.9.23
flake8==2.1.0
kombu==3.0.30
lxml==3.4.4
mccabe==0.3.1
pep8==1.6.2
Pillow==2.7.0
pyflakes==0.9.2
pymongo==2.7.1
pysolr==3.2.0
python-magic==0.4.6
pytz==2015.4
qrcode==5.0.1
redis==2.10.3
requests==2.3.0
Serializer==0.2.1
simplejson==3.6.5
six==1.9.0
sorl-thumbnail==11.12
South==1.0.2
tld==0.7.2
uWSGI==2.0.15
wheel==0.24.0
thrift==0.9.2
chardet==2.3.0
numpy-stl==1.7.1
pycrypto==2.6.1
raven==5.10.1
pymesh==1.0.2
MySQL-python==1.2.5
msgpack-python==0.4.7
tornado==4.3
line-profiler==1.0
oss2==2.2.0
xlrd==1.0.0
xlwt==1.1.2
mongoengine==0.14.3
eventlet==0.21.0
flask==0.12.2
python-socketio==1.8.4
